Fishes change colour correctly
The walls are clearer than before, the room is bigger too. I could touch the sky and its stars outside if I wasn't so afraid to. I fear the cold and empty twinkling of the astral bodies that encroach ever closer. I don't want to become a star, not yet, its not time, I still have work to do.

Like the title says, the fish change colour correctly and match their respective numbers. I also got rid of the mouse press because we won't need them. Oh and I changed some text to reflect the button press instead of a mouse press